xemed_agent/requirements.txt
# XEMED Agent Requirements

# Core dependencies
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.1.0
langchain-community>=0.0.20

# Optional: for better JSON parsing
pydantic>=2.0.0

